+++ py-src/data_formulator/app.py
@@ -10,10 +10,8 @@ mimetypes.add_type('application/javascript', '.js')
 mimetypes.add_type('application/javascript', '.mjs')
 
 import flask
-from flask import Flask, request, send_from_directory, has_request_context, redirect, url_for, session, jsonify
+from flask import Flask, request, send_from_directory, session
 from flask import stream_with_context, Response
-import html
-import pandas as pd
 
 import webbrowser
 import threading
@@ -26,33 +24,23 @@ from pathlib import Path
 
 from vega_datasets import data as vega_data
 
-from data_formulator.agents.agent_concept_derive import ConceptDeriveAgent
-from data_formulator.agents.agent_py_concept_derive import PyConceptDeriveAgent
-
-from data_formulator.agents.agent_py_data_transform import PythonDataTransformationAgent
-from data_formulator.agents.agent_sql_data_transform import SQLDataTransformationAgent
-from data_formulator.agents.agent_py_data_rec import PythonDataRecAgent
-from data_formulator.agents.agent_sql_data_rec import SQLDataRecAgent
-
-from data_formulator.agents.agent_sort_data import SortDataAgent
-from data_formulator.agents.agent_data_load import DataLoadAgent
-from data_formulator.agents.agent_data_clean import DataCleanAgent
-from data_formulator.agents.agent_code_explanation import CodeExplanationAgent
-
-from data_formulator.agents.client_utils import Client
-
 from dotenv import load_dotenv
 import secrets
-
-from data_formulator.db_manager import db_manager
 
 APP_ROOT = Path(os.path.join(Path(__file__).parent)).absolute()
 
 import os
-import tempfile
+
+# blueprints
+from data_formulator.tables_routes import tables_bp
+from data_formulator.agent_routes import agent_bp
 
 app = Flask(__name__, static_url_path='', static_folder=os.path.join(APP_ROOT, "dist"))
 app.secret_key = secrets.token_hex(16)  # Generate a random secret key for sessions
+
+# register blueprints
+app.register_blueprint(tables_bp)
+app.register_blueprint(agent_bp)
 
 print(APP_ROOT)
 
@@ -80,18 +68,6 @@ for handler in logging.getLogger().handlers:
 logger.info("Application level log")  # General application logging
 app.logger.info("Flask specific log") # Web request related logging
 
-def get_client(model_config):
-    for key in model_config:
-        model_config[key] = model_config[key].strip()
-
-    client = Client(
-        model_config["endpoint"],
-        model_config["model"],
-        model_config["api_key"] if "api_key" in model_config else None,
-        html.escape(model_config["api_base"]) if "api_base" in model_config else None,
-        model_config["api_version"] if "api_version" in model_config else None)
-
-    return client
 
 @app.route('/api/vega-datasets')
 def get_example_dataset_list():
@@ -156,103 +132,6 @@ def get_datasets(path):
     response = data_object
     return response
 
-@app.route('/api/check-available-models', methods=['GET', 'POST'])
-def check_available_models():
-    results = []
-    
-    # Define configurations for different providers
-    providers = ['openai', 'azure', 'anthropic', 'gemini', 'ollama']
-
-    for provider in providers:
-        # Skip if provider is not enabled
-        if not os.getenv(f"{provider.upper()}_ENABLED", "").lower() == "true":
-            continue
-        
-        api_key = os.getenv(f"{provider.upper()}_API_KEY", "")
-        api_base = os.getenv(f"{provider.upper()}_API_BASE", "")
-        api_version = os.getenv(f"{provider.upper()}_API_VERSION", "")
-        models = os.getenv(f"{provider.upper()}_MODELS", "")
-
-        if not (api_key or api_base):
-            continue
-
-        if not models:
-            continue
-
-        # Build config for each model
-        for model in models.split(","):
-            model = model.strip()
-            if not model:
-                continue
-
-            model_config = {
-                "id": f"{provider}-{model}-{api_key}-{api_base}-{api_version}",
-                "endpoint": provider,
-                "model": model,
-                "api_key": api_key,
-                "api_base": api_base,
-                "api_version": api_version
-            }
-            
-            try:
-                client = get_client(model_config)
-                response = client.get_completion(
-                    messages=[
-                        {"role": "system", "content": "You are a helpful assistant."},
-                        {"role": "user", "content": "Respond 'I can hear you.' if you can hear me."},
-                    ]
-                )
-                
-                if "I can hear you." in response.choices[0].message.content:
-                    results.append(model_config)
-            except Exception as e:
-                print(f"Error testing {provider} model {model}: {e}")
-                
-    return json.dumps(results)
-
-@app.route('/api/test-model', methods=['GET', 'POST'])
-def test_model():
-    
-    if request.is_json:
-        app.logger.info("# code query: ")
-        content = request.get_json()
-
-        # contains endpoint, key, model, api_base, api_version
-        logger.info("content------------------------------")
-        logger.info(content)
-
-        client = get_client(content['model'])
-
-        try:
-            response = client.get_completion(
-                messages=[
-                    {"role": "system", "content": "You are a helpful assistant."},
-                    {"role": "user", "content": "Respond 'I can hear you.' if you can hear me. Do not say anything other than 'I can hear you.'"},
-                ]
-            )
-
-            logger.info(f"model: {content['model']}")
-            logger.info(f"welcome message: {response.choices[0].message.content}")
-
-            if "I can hear you." in response.choices[0].message.content:
-                result = {
-                    "model": content['model'],
-                    "status": 'ok',
-                    "message": ""
-                }
-        except Exception as e:
-            logger.info(f"Error: {e}")
-            error_message = str(e)
-            result = {
-                "model": content['model'],
-                "status": 'error',
-                "message": error_message,
-            }
-    else:
-        result = {'status': 'error'}
-    
-    return json.dumps(result)
-
 @app.route("/", defaults={"path": ""})
 def index_alt(path):
     logger.info(app.static_folder)
@@ -310,279 +189,6 @@ def streamed_response():
             yield json.dumps(spec)
     return Response(stream_with_context(generate()))
 
-
-###### agent related functions ######
-
-@app.route('/api/process-data-on-load', methods=['GET', 'POST'])
-def process_data_on_load_request():
-
-    if request.is_json:
-        app.logger.info("# process data query: ")
-        content = request.get_json()
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        app.logger.info(f" model: {content['model']}")
-        
-        agent = DataLoadAgent(client=client)
-        candidates = agent.run(content["input_data"])
-        
-        candidates = [c['content'] for c in candidates if c['status'] == 'ok']
-
-        response = flask.jsonify({ "status": "ok", "token": token, "result": candidates })
-    else:
-        response = flask.jsonify({ "token": -1, "status": "error", "result": [] })
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route('/api/derive-concept-request', methods=['GET', 'POST'])
-def derive_concept_request():
-
-    if request.is_json:
-        app.logger.info("# code query: ")
-        content = request.get_json()
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        app.logger.info(f" model: {content['model']}")
-        agent = ConceptDeriveAgent(client=client)
-
-        #print(content["input_data"])
-
-        candidates = agent.run(content["input_data"], [f['name'] for f in content["input_fields"]], 
-                                       content["output_name"], content["description"])
-        
-        candidates = [c['code'] for c in candidates if c['status'] == 'ok']
-
-        response = flask.jsonify({ "status": "ok", "token": token, "result": candidates })
-    else:
-        response = flask.jsonify({ "token": -1, "status": "error", "result": [] })
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route('/api/derive-py-concept', methods=['GET', 'POST'])
-def derive_py_concept():
-
-    if request.is_json:
-        app.logger.info("# code query: ")
-        content = request.get_json()
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        app.logger.info(f" model: {content['model']}")
-        agent = PyConceptDeriveAgent(client=client)
-
-        #print(content["input_data"])
-
-        results = agent.run(content["input_data"], [f['name'] for f in content["input_fields"]], 
-                                       content["output_name"], content["description"])
-        
-        response = flask.jsonify({ "status": "ok", "token": token, "results": results })
-    else:
-        response = flask.jsonify({ "token": -1, "status": "error", "results": [] })
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-@app.route('/api/clean-data', methods=['GET', 'POST'])
-def clean_data_request():
-
-    if request.is_json:
-        app.logger.info("# data clean request")
-        content = request.get_json()
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        app.logger.info(f" model: {content['model']}")
-        
-        agent = DataCleanAgent(client=client)
-
-        candidates = agent.run(content['content_type'], content["raw_data"], content["image_cleaning_instruction"])
-        
-        candidates = [c for c in candidates if c['status'] == 'ok']
-
-        response = flask.jsonify({ "status": "ok", "token": token, "result": candidates })
-    else:
-        response = flask.jsonify({ "token": -1, "status": "error", "result": [] })
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route('/api/codex-sort-request', methods=['GET', 'POST'])
-def sort_data_request():
-
-    if request.is_json:
-        app.logger.info("# sort query: ")
-        content = request.get_json()
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        agent = SortDataAgent(client=client)
-        candidates = agent.run(content['field'], content['items'])
-
-        #candidates, dialog = limbo_concept.call_codex_sort(content["items"], content["field"])
-        candidates = candidates if candidates != None else []
-        response = flask.jsonify({ "status": "ok", "token": token, "result": candidates })
-    else:
-        response = flask.jsonify({ "token": -1, "status": "error", "result": [] })
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-@app.route('/api/derive-data', methods=['GET', 'POST'])
-def derive_data():
-
-    if request.is_json:
-        app.logger.info("# request data: ")
-        content = request.get_json()        
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        # each table is a dict with {"name": xxx, "rows": [...]}
-        input_tables = content["input_tables"]
-        new_fields = content["new_fields"]
-        instruction = content["extra_prompt"]
-        language = content.get("language", "python") # whether to use sql or python, default to python
-        
-        max_repair_attempts = content["max_repair_attempts"] if "max_repair_attempts" in content else 1
-
-        if "additional_messages" in content:
-            prev_messages = content["additional_messages"]
-        else:
-            prev_messages = []
-
-        logger.info("== input tables ===>")
-        for table in input_tables:
-            logger.info(f"===> Table: {table['name']} (first 5 rows)")
-            logger.info(table['rows'][:5])
-
-        logger.info("== user spec ===")
-        logger.info(new_fields)
-        logger.info(instruction)
-
-        mode = "transform"
-        if len(new_fields) == 0:
-            mode = "recommendation"
-
-        conn = db_manager.get_connection(session['session_id']) if language == "sql" else None
-
-        if mode == "recommendation":
-            # now it's in recommendation mode
-            agent = SQLDataRecAgent(client=client, conn=conn) if language == "sql" else PythonDataRecAgent(client=client)
-            results = agent.run(input_tables, instruction)
-        else:
-            agent = SQLDataTransformationAgent(client=client, conn=conn) if language == "sql" else PythonDataTransformationAgent(client=client)
-            results = agent.run(input_tables, instruction, [field['name'] for field in new_fields], prev_messages)
-
-        repair_attempts = 0
-        while results[0]['status'] == 'error' and repair_attempts < max_repair_attempts: # try up to n times
-            error_message = results[0]['content']
-            new_instruction = f"We run into the following problem executing the code, please fix it:\n\n{error_message}\n\nPlease think step by step, reflect why the error happens and fix the code so that no more errors would occur."
-
-            prev_dialog = results[0]['dialog']
-
-            if mode == "transform":
-                results = agent.followup(input_tables, prev_dialog, [field['name'] for field in new_fields], new_instruction)
-            if mode == "recommendation":
-                results = agent.followup(input_tables, prev_dialog, new_instruction)
-
-            repair_attempts += 1
-        
-        if conn:
-            conn.close()
-        
-        response = flask.jsonify({ "token": token, "status": "ok", "results": results })
-    else:
-        response = flask.jsonify({ "token": "", "status": "error", "results": [] })
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-@app.route('/api/refine-data', methods=['GET', 'POST'])
-def refine_data():
-
-    if request.is_json:
-        app.logger.info("# request data: ")
-        content = request.get_json()        
-        token = content["token"]
-
-
-        client = get_client(content['model'])
-
-        # each table is a dict with {"name": xxx, "rows": [...]}
-        input_tables = content["input_tables"]
-        output_fields = content["output_fields"]
-        dialog = content["dialog"]
-        new_instruction = content["new_instruction"]
-        max_repair_attempts = content.get("max_repair_attempts", 1)
-        language = content.get("language", "python") # whether to use sql or python, default to python
-
-        logger.info("== input tables ===>")
-        for table in input_tables:
-            logger.info(f"===> Table: {table['name']} (first 5 rows)")
-            logger.info(table['rows'][:5])
-        
-        logger.info("== user spec ===>")
-        logger.info(output_fields)
-        logger.info(new_instruction)
-
-        conn = db_manager.get_connection(session['session_id']) if language == "sql" else None
-
-        # always resort to the data transform agent       
-        agent = SQLDataTransformationAgent(client=client, conn=conn) if language == "sql" else PythonDataTransformationAgent(client=client)
-        results = agent.followup(input_tables, dialog, [field['name'] for field in output_fields], new_instruction)
-
-        repair_attempts = 0
-        while results[0]['status'] == 'error' and repair_attempts < max_repair_attempts: # only try once
-            error_message = results[0]['content']
-            new_instruction = f"We run into the following problem executing the code, please fix it:\n\n{error_message}\n\nPlease think step by step, reflect why the error happens and fix the code so that no more errors would occur."
-            prev_dialog = results[0]['dialog']
-
-            results = agent.followup(input_tables, prev_dialog, [field['name'] for field in output_fields], new_instruction)
-            repair_attempts += 1
-
-        if conn:
-            conn.close()
-
-        response = flask.jsonify({ "token": token, "status": "ok", "results": results})
-    else:
-        response = flask.jsonify({ "token": "", "status": "error", "results": []})
-
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-
-@app.route('/api/code-expl', methods=['GET', 'POST'])
-def request_code_expl():
-    if request.is_json:
-        app.logger.info("# request data: ")
-        content = request.get_json()        
-        token = content["token"]
-
-        client = get_client(content['model'])
-
-        # each table is a dict with {"name": xxx, "rows": [...]}
-        input_tables = content["input_tables"]
-        code = content["code"]
-        
-        code_expl_agent = CodeExplanationAgent(client=client)
-        expl = code_expl_agent.run(input_tables, code)
-    else:
-        expl = ""
-    return expl
-
 @app.route('/api/get-session-id', methods=['GET', 'POST'])
 def get_session_id():
     """Endpoint to get or confirm a session ID from the client"""
@@ -609,7 +215,6 @@ def get_session_id():
         "status": "ok",
         "session_id": session['session_id']
     })
-    
 
 @app.route('/api/app-config', methods=['GET'])
 def get_app_config():
@@ -620,583 +225,6 @@ def get_app_config():
         "SESSION_ID": session.get('session_id', None)
     }
     return flask.jsonify(config)
-
-
-@app.route('/api/tables', methods=['GET'])
-def list_tables():
-    """List all tables in the current session"""
-    try:
-        result = []
-        with db_manager.connection(session['session_id']) as db:
-            tables = db.execute("SHOW TABLES").fetchall()
-        
-            for table in tables:
-                table_name = table[0]
-                # Get column information
-                columns = db.execute(f"DESCRIBE {table_name}").fetchall()
-                # Get row count
-                row_count = db.execute(f"SELECT COUNT(*) FROM {table_name}").fetchone()[0]
-                sample_rows = db.execute(f"SELECT * FROM {table_name} LIMIT 1000").fetchall()
-                
-                # Check if this is a view or a table
-                try:
-                    # Get both view existence and source in one query
-                    view_info = db.execute(f"SELECT view_name, sql FROM duckdb_views() WHERE view_name = '{table_name}'").fetchone()
-                    view_source = view_info[1] if view_info else None
-                except Exception as e:
-                    # If the query fails, assume it's a regular table
-                    view_source = None
-
-                result.append({
-                    "name": table_name,
-                    "columns": [{"name": col[0], "type": col[1]} for col in columns],
-                    "row_count": row_count,
-                    "sample_rows": [dict(zip([col[0] for col in columns], row)) for row in sample_rows],
-                    "view_source": view_source
-                })
-        
-        return jsonify({
-            "status": "success",
-            "tables": result
-        })
-    except Exception as e:
-        print(e)
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
-
-def assemble_query(aggregate_fields_and_functions, group_fields, columns, table_name):
-    """
-    Assembles a SELECT query string based on binning, aggregation, and grouping specifications.
-    
-    Args:
-        bin_fields (list): Fields to be binned into ranges
-        aggregate_fields_and_functions (list): List of tuples (field, function) for aggregation
-        group_fields (list): Fields to group by
-        columns (list): All available column names
-    
-    Returns:
-        str: The assembled SELECT query projection part
-    """
-    select_parts = []
-    output_column_names = []
-
-    # Handle aggregate fields and functions
-    for field, function in aggregate_fields_and_functions:
-        if field is None:
-            # Handle count(*) case
-            if function.lower() == 'count':
-                select_parts.append('COUNT(*) as _count')
-                output_column_names.append('_count')
-        elif field in columns:
-            if function.lower() == 'count':
-                alias = f'_count'
-                select_parts.append(f'COUNT(*) as {alias}')
-                output_column_names.append(alias)
-            else:
-                # Sanitize function name and create alias
-                if function in ["avg", "average", "mean"]:
-                    aggregate_function = "AVG"
-                else:
-                    aggregate_function = function.upper()
-                
-                alias = f'{field}_{function}'
-                select_parts.append(f'{aggregate_function}("{field}") as {alias}')
-                output_column_names.append(alias)
-
-    # Handle group fields
-    for field in group_fields:
-        if field in columns:
-            select_parts.append(f'"{field}"')
-            output_column_names.append(field)
-    # If no fields are specified, select all columns
-    if not select_parts:
-        select_parts = ["*"]
-        output_column_names = columns
-
-    from_clause = f"FROM {table_name}"
-    group_by_clause = f"GROUP BY {', '.join(group_fields)}" if len(group_fields) > 0 and len(aggregate_fields_and_functions) > 0 else ""
-
-    query = f"SELECT {', '.join(select_parts)} {from_clause} {group_by_clause}"
-    return query, output_column_names
-
-@app.route('/api/tables/sample-table', methods=['POST'])
-def sample_table():
-    """Sample a table"""
-    try:
-        data = request.get_json()
-        table_id = data.get('table')
-        sample_size = data.get('size', 1000)
-        aggregate_fields_and_functions = data.get('aggregate_fields_and_functions', []) # each element is a tuple (field, function)
-        select_fields = data.get('select_fields', []) # if empty, we want to include all fields
-        method = data.get('method', 'random') # one of 'random', 'head', 'bottom'
-        order_by_fields = data.get('order_by_fields', [])
-
-        print(f"sample_table: {table_id}, {sample_size}, {aggregate_fields_and_functions}, {select_fields}, {method}, {order_by_fields}")
-        
-        total_row_count = 0
-        # Validate field names against table columns to prevent SQL injection
-        with db_manager.connection(session['session_id']) as db:
-            # Get valid column names
-            columns = [col[0] for col in db.execute(f"DESCRIBE {table_id}").fetchall()]
-            
-            # Filter order_by_fields to only include valid column names
-            valid_order_by_fields = [field for field in order_by_fields if field in columns]
-            valid_aggregate_fields_and_functions = [
-                field_and_function for field_and_function in aggregate_fields_and_functions 
-                if field_and_function[0] is None or field_and_function[0] in columns
-            ]
-            valid_select_fields = [field for field in select_fields if field in columns]
-
-            query, output_column_names = assemble_query(valid_aggregate_fields_and_functions, valid_select_fields, columns, table_id)
-
-            # Modify the original query to include the count:
-            count_query = f"SELECT *, COUNT(*) OVER () as total_count FROM ({query}) as subq LIMIT 1"
-            result = db.execute(count_query).fetchone()
-            total_row_count = result[-1] if result else 0
-
-            # Add ordering and limit to the main query
-            if method == 'random':
-                query += f" ORDER BY RANDOM() LIMIT {sample_size}"
-            elif method == 'head':
-                if valid_order_by_fields:
-                    # Build ORDER BY clause with validated fields
-                    order_by_clause = ", ".join([f'"{field}"' for field in valid_order_by_fields])
-                    query += f" ORDER BY {order_by_clause} LIMIT {sample_size}"
-                else:
-                    query += f" LIMIT {sample_size}"
-            elif method == 'bottom':
-                if valid_order_by_fields:
-                    # Build ORDER BY clause with validated fields in descending order
-                    order_by_clause = ", ".join([f'"{field}" DESC' for field in valid_order_by_fields])
-                    query += f" ORDER BY {order_by_clause} LIMIT {sample_size}"
-                else:
-                    query += f" ORDER BY ROWID DESC LIMIT {sample_size}"
-
-            result = db.execute(query).fetchall()
-
-        return jsonify({
-            "status": "success",
-            "rows": [dict(zip(output_column_names, row)) for row in result],
-            "total_row_count": total_row_count
-        })
-    except Exception as e:
-        print(e)
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
-
-@app.route('/api/tables/get-table', methods=['GET'])
-def get_table_data():
-    """Get data from a specific table"""
-    try:
-        with db_manager.connection(session['session_id']) as db:
-
-            table_name = request.args.get('table_name')
-            # Get pagination parameters
-            page = int(request.args.get('page', 1))
-            page_size = int(request.args.get('page_size', 100))
-            offset = (page - 1) * page_size
-            
-            if not table_name:
-                return jsonify({
-                    "status": "error",
-                    "message": "Table name is required"
-                }), 400
-            
-            # Get total count
-            total_rows = db.execute(f"SELECT COUNT(*) FROM {table_name}").fetchone()[0]
-            
-            # Get paginated data
-            result = db.execute(
-                f"SELECT * FROM {table_name} LIMIT {page_size} OFFSET {offset}"
-            ).fetchall()
-            
-            # Get column names
-            columns = [col[0] for col in db.execute(f"DESCRIBE {table_name}").fetchall()]
-            
-            # Convert to list of dictionaries
-            rows = [dict(zip(columns, row)) for row in result]
-        
-            return jsonify({
-                "status": "success",
-                "table_name": table_name,
-                "columns": columns,
-                "rows": rows,
-                "total_rows": total_rows,
-                "page": page,
-                "page_size": page_size
-            })
-    
-    except Exception as e:
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
-
-@app.route('/api/tables/create-table', methods=['POST'])
-def create_table():
-    """Create a new table from uploaded data"""
-    try:
-        if 'file' not in request.files:
-            return jsonify({"status": "error", "message": "No file provided"}), 400
-        
-        file = request.files['file']
-        table_name = request.form.get('table_name')
-
-        print(f"table_name: {table_name}")
-        print(f"file: {file.filename}")
-        print(f"file: {file}")
-        
-        if not table_name:
-            return jsonify({"status": "error", "message": "No table name provided"}), 400
-            
-        # Sanitize table name:
-        # 1. Convert to lowercase
-        # 2. Replace hyphens with underscores
-        # 3. Replace spaces with underscores
-        # 4. Remove any other special characters
-        sanitized_table_name = table_name.lower()
-        sanitized_table_name = sanitized_table_name.replace('-', '_')
-        sanitized_table_name = sanitized_table_name.replace(' ', '_')
-        sanitized_table_name = ''.join(c for c in sanitized_table_name if c.isalnum() or c == '_')
-        
-        # Ensure table name starts with a letter
-        if not sanitized_table_name or not sanitized_table_name[0].isalpha():
-            sanitized_table_name = 'table_' + sanitized_table_name
-            
-        # Verify we have a valid table name after sanitization
-        if not sanitized_table_name:
-            return jsonify({"status": "error", "message": "Invalid table name"}), 400
-            
-        with db_manager.connection(session['session_id']) as db:
-        
-            # Read file based on extension
-            if file.filename.endswith('.csv'):
-                df = pd.read_csv(file)
-            elif file.filename.endswith(('.xlsx', '.xls')):
-                df = pd.read_excel(file)
-            elif file.filename.endswith('.json'):
-                df = pd.read_json(file)
-            else:
-                return jsonify({"status": "error", "message": "Unsupported file format"}), 400
-            
-            # Create table
-            db.register('df_temp', df)
-            db.execute(f"CREATE TABLE {sanitized_table_name} AS SELECT * FROM df_temp")
-            db.execute("DROP VIEW df_temp")  # Drop the temporary view after creating the table
-            
-            return jsonify({
-                "status": "success",
-                "table_name": sanitized_table_name,
-                "row_count": len(df),
-                "columns": list(df.columns)
-            })
-    
-    except Exception as e:
-        print(e)
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
-
-@app.route('/api/tables/delete-table', methods=['POST'])
-def drop_table():
-    """Drop a table or view"""
-    try:
-        data = request.get_json()
-        table_name = data.get('table_name')
-        
-        if not table_name:
-            return jsonify({"status": "error", "message": "No table name provided"}), 400
-            
-        with db_manager.connection(session['session_id']) as db:
-            # First try to drop it as a view
-            db.execute(f"DROP VIEW IF EXISTS {table_name}")
-            # Then try to drop it as a table
-            db.execute(f"DROP TABLE IF EXISTS {table_name}")
-        
-            return jsonify({
-                "status": "success",
-                "message": f"Table/view {table_name} dropped"
-            })
-    
-    except Exception as e:
-        print(e)
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
-
-
-@app.route('/api/tables/upload-db-file', methods=['POST'])
-def upload_db_file():
-    """Upload a db file"""
-    try:
-        if 'file' not in request.files:
-            return jsonify({"status": "error", "message": "No file provided"}), 400
-        
-        file = request.files['file']
-        if not file.filename.endswith('.db'):
-            return jsonify({"status": "error", "message": "Invalid file format. Only .db files are supported"}), 400
-
-        # Get the session ID
-        if 'session_id' not in session:
-            return jsonify({"status": "error", "message": "No session ID found"}), 400
-        
-        session_id = session['session_id']
-        
-        # Create temp directory if it doesn't exist
-        temp_dir = os.path.join(tempfile.gettempdir())
-        os.makedirs(temp_dir, exist_ok=True)
-        
-        # Save the file temporarily to verify it
-        temp_db_path = os.path.join(temp_dir, f"temp_{session_id}.db")
-        file.save(temp_db_path)
-        
-        # Verify if it's a valid DuckDB file
-        try:
-            import duckdb
-            # Try to connect to the database
-            conn = duckdb.connect(temp_db_path, read_only=True)
-            # Try a simple query to verify it's a valid database
-            conn.execute("SELECT 1").fetchall()
-            conn.close()
-            
-            # If we get here, the file is valid - move it to final location
-            db_file_path = os.path.join(temp_dir, f"df_{session_id}.db")
-            os.replace(temp_db_path, db_file_path)
-            
-            # Update the db_manager's file mapping
-            db_manager._db_files[session_id] = db_file_path
-            
-        except Exception as db_error:
-            # Clean up temp file
-            if os.path.exists(temp_db_path):
-                os.remove(temp_db_path)
-            return jsonify({
-                "status": "error",
-                "message": f"Invalid DuckDB database file: {str(db_error)}"
-            }), 400
-        
-        return jsonify({
-            "status": "success",
-            "message": "Database file uploaded successfully",
-            "session_id": session_id
-        })
-        
-    except Exception as e:
-        logger.error(f"Error uploading database file: {str(e)}")
-        return jsonify({
-            "status": "error", 
-            "message": f"Failed to upload database file: {str(e)}"
-        }), 500
-
-@app.route('/api/tables/download-db-file', methods=['GET'])
-def download_db_file():
-    """Download the db file for a session"""
-    try:
-        # Check if session exists
-        if 'session_id' not in session:
-            return jsonify({
-                "status": "error",
-                "message": "No session ID found"
-            }), 400
-        
-        session_id = session['session_id']
-        
-        # Get the database file path from db_manager
-        if session_id not in db_manager._db_files:
-            return jsonify({
-                "status": "error",
-                "message": "No database file found for this session"
-            }), 404
-            
-        db_file_path = db_manager._db_files[session_id]
-        
-        # Check if file exists
-        if not os.path.exists(db_file_path):
-            return jsonify({
-                "status": "error",
-                "message": "Database file not found"
-            }), 404
-            
-        # Generate a filename for download
-        download_name = f"data_formulator_{session_id}.db"
-        
-        # Return the file as an attachment
-        return send_from_directory(
-            os.path.dirname(db_file_path),
-            os.path.basename(db_file_path),
-            as_attachment=True,
-            download_name=download_name,
-            mimetype='application/x-sqlite3'
-        )
-        
-    except Exception as e:
-        logger.error(f"Error downloading database file: {str(e)}")
-        return jsonify({
-            "status": "error",
-            "message": f"Failed to download database file: {str(e)}"
-        }), 500
-
-
-@app.route('/api/tables/reset-db-file', methods=['POST'])
-def reset_db_file():
-    """Reset the db file for a session"""
-    try:
-        if 'session_id' not in session:
-            return jsonify({
-                "status": "error",
-                "message": "No session ID found"
-            }), 400
-            
-        session_id = session['session_id']
-
-        print(f"session_id: {session_id}")
-        
-        # First check if there's a reference in db_manager
-        if session_id in db_manager._db_files:
-            db_file_path = db_manager._db_files[session_id]
-            
-            # Remove the file if it exists
-            if db_file_path and os.path.exists(db_file_path):
-                os.remove(db_file_path)
-            
-            # Clear the reference
-            db_manager._db_files[session_id] = None
-            
-        # Also check for any temporary files
-        temp_db_path = os.path.join(tempfile.gettempdir(), f"temp_{session_id}.db")
-        if os.path.exists(temp_db_path):
-            os.remove(temp_db_path)
-            
-        # Check for the main db file
-        main_db_path = os.path.join(tempfile.gettempdir(), f"df_{session_id}.db")
-        if os.path.exists(main_db_path):
-            os.remove(main_db_path)
-
-        return jsonify({
-            "status": "success",
-            "message": "Database file reset successfully"
-        })
-
-    except Exception as e:
-        logger.error(f"Error resetting database file: {str(e)}")
-        return jsonify({
-            "status": "error",
-            "message": f"Failed to reset database file: {str(e)}"
-        }), 500
-
-
-@app.route('/api/tables/query', methods=['POST'])
-def query_table():
-    """Execute a query on a table"""
-    try:
-        data = request.get_json()
-
-        query = data.get('query')
-
-        if not query:
-            return jsonify({"status": "error", "message": "No query provided"}), 400
-        
-        with db_manager.connection(session['session_id']) as db:
-            result = db.execute(query).fetch_df()
-        
-            return jsonify({
-                "status": "success",
-                "rows": result.to_dict('records'),
-                "columns": list(result.columns)
-            })
-    
-    except Exception as e:
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
-
-
-# Example of a more complex query endpoint
-@app.route('/api/tables/analyze', methods=['POST'])
-def analyze_table():
-    """Get basic statistics about a table"""
-    try:
-        data = request.get_json()
-        table_name = data.get('table_name')
-        
-        if not table_name:
-            return jsonify({"status": "error", "message": "No table name provided"}), 400
-        
-        with db_manager.connection(session['session_id']) as db:
-        
-            # Get column information
-            columns = db.execute(f"DESCRIBE {table_name}").fetchall()
-            
-            print(f"columns: {columns}")
-            stats = []
-            for col in columns:
-                col_name = col[0]
-                col_type = col[1]
-                
-                # Properly quote column names to avoid SQL keywords issues
-                quoted_col_name = f'"{col_name}"'
-                
-                # Basic stats query
-                stats_query = f"""
-                SELECT 
-                    COUNT(*) as count,
-                    COUNT(DISTINCT {quoted_col_name}) as unique_count,
-                    COUNT(*) - COUNT({quoted_col_name}) as null_count
-                FROM {table_name}
-                """
-                
-                # Add numeric stats if applicable
-                if col_type in ['INTEGER', 'DOUBLE', 'DECIMAL']:
-                    stats_query = f"""
-                    SELECT 
-                        COUNT(*) as count,
-                        COUNT(DISTINCT {quoted_col_name}) as unique_count,
-                        COUNT(*) - COUNT({quoted_col_name}) as null_count,
-                        MIN({quoted_col_name}) as min_value,
-                        MAX({quoted_col_name}) as max_value,
-                        AVG({quoted_col_name}) as avg_value
-                    FROM {table_name}
-                    """
-                
-                col_stats = db.execute(stats_query).fetchone()
-                
-                # Create a dictionary with appropriate keys based on column type
-                if col_type in ['INTEGER', 'DOUBLE', 'DECIMAL']:
-                    stats_dict = dict(zip(
-                        ["count", "unique_count", "null_count", "min", "max", "avg"],
-                        col_stats
-                    ))
-                else:
-                    stats_dict = dict(zip(
-                        ["count", "unique_count", "null_count"],
-                        col_stats
-                    ))
-                
-                stats.append({
-                    "column": col_name,
-                    "type": col_type,
-                    "statistics": stats_dict
-                })
-        
-        return jsonify({
-            "status": "success",
-            "table_name": table_name,
-            "statistics": stats
-        })
-    
-    except Exception as e:
-        print(e)
-        return jsonify({
-            "status": "error",
-            "message": str(e)
-        }), 500
 
 
 def parse_args() -> argparse.Namespace:
